backend/requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-perms==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
future==0.18.2
gunicorn==19.9.0
idna==2.10
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
joblib==1.0.1
livereload==2.6.3
lunr==0.5.8
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
mkdocs==1.1.2
nltk==3.6.2
oauthlib==3.1.1
packaging==20.9
Pillow==8.4.0
pluggy==0.13.1
psycopg2==2.9
psycopg2-binary==2.9.2
py==1.10.0
pycparser==2.21
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.26.0
requests-oauthlib==1.3.0
Serializer==0.2.1
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
tqdm==4.61.0
uritemplate==4.1.1
urllib3==1.26.7
